% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alternativeROC-common-args.R
\name{alternativeROC-common-args}
\alias{alternativeROC-common-args}
\title{Arguments used across the functions of the alternativeROC package.}
\arguments{
\item{roc}{Object of class pROC.}

\item{se}{Sensitivity.}

\item{sp}{Specificity.}

\item{ppv}{Positive predictive value.}

\item{npv}{Negative predictive value.}

\item{prevalence}{Prevalence of the endpoint in the study population.}

\item{boot.n}{Number of bootstrap replicates. Default: 2000.}

\item{quantiles}{Quantiles. Default: c(0.5,.025,.975).}

\item{conf.level}{Width of the confidence interval. Default: 0.95 (i.e., 95\% CI).}
}
\value{
No return value, used for the documentation of
the functions of the package.
}
\description{
Arguments used across the functions of the alternativeROC package.
}
