% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alternativeROC.R
\docType{package}
\name{alternativeROC-package}
\alias{alternativeROC-package}
\title{alternativeROC: Alternative and Fast ROC Analysis}
\description{
Alternative and fast algorithms for the analysis of receiver operating characteristics curves (ROC curves) as described in Thomas et al. (2017) \doi{10.1186/s41512-017-0017-y} and Thomas et al. (2023) \doi{10.1016/j.ajogmf.2023.101110}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://bitbucket.org/SQ4/alternativeROC}
  \item Report bugs at \url{https://bitbucket.org/SQ4/alternativeROC/issues}
}

}
\author{
\strong{Maintainer}: Gregoire Thomas \email{gregoire.thomas@SQU4RE.com} (\href{https://orcid.org/0000-0002-6247-9438}{ORCID})

Other contributors:
\itemize{
  \item Robin Tuytten \email{robin.tuytten@metabolomicdiagnostics.com} (\href{https://orcid.org/0000-0002-8734-7335}{ORCID}) [contributor]
  \item Jef Moerman [contributor]
  \item Xavier Robin (Co-author of delong.cpp from package pROC) [copyright holder]
  \item Stefan Siegert (Co-author of delong.cpp from package pROC) [copyright holder]
}

}
\keyword{internal}
