% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alternativeROC.R
\name{alternativeROC}
\alias{alternativeROC}
\title{alternativeROC}
\description{
Alternative analysis methods for the analysis of receiver operating characteristics curve (ROC curve) as described in Thomas et al. (2017) <doi:10.1186/s41512-017-0017-y>.
}
\author{
Gregoire Thomas <gregoire.thomas@SQU4RE.com>
}
