% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-roc.R
\name{npv.seatsp}
\alias{npv.seatsp}
\title{Diagnostic performance: Sensitivity from specificity, NPV and prevalence}
\usage{
npv.seatsp(sp, npv, prevalence)
}
\arguments{
\item{sp}{Specificity.}

\item{npv}{Negative predictive value.}

\item{prevalence}{Prevalence of the endpoint in the study population.}
}
\value{
Sensitivity from specificity, NPV and prevalence.
}
\description{
Diagnostic performance: Sensitivity from specificity, negative predictive value (NPV) and prevalence of the endpoint
}
