% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-roc.R
\name{npv.spatse}
\alias{npv.spatse}
\title{Diagnostic performance: Specificity from sensitivity, NPV and prevalence}
\usage{
npv.spatse(se, npv, prevalence)
}
\arguments{
\item{se}{Sensitivity.}

\item{npv}{Negative predictive value.}

\item{prevalence}{Prevalence of the endpoint in the study population.}
}
\value{
Specificity from sensitivity, NPV and prevalence.
}
\description{
Diagnostic performance: Specificity from sensitivity, negative predictive value (NPV) and prevalence of the endpoint
}
