% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p.auc.R
\name{p.auc}
\alias{p.auc}
\title{p[AUC=0.5]}
\usage{
p.auc(roc, ref = 0.5)
}
\arguments{
\item{roc}{Object of class pROC.}

\item{ref}{Reference AUC. Default 0.5.}
}
\value{
p value.
}
\description{
p[AUC=0.5] using DeLong's methods (DeLong et al. 1988).
}
