% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-roc.R
\name{ppv}
\alias{ppv}
\title{Diagnostic performance: Positive predictive values from a ROC curve}
\usage{
ppv(roc, prevalence)
}
\arguments{
\item{roc}{Object of class pROC.}

\item{prevalence}{Prevalence of the endpoint in the study population.}
}
\value{
A matrix with the following columns:
\itemize{
\item{\code{threshold}: The thresholds used to compute the sensitivity and specificity.}
\item{\code{sensitivity}: The sensitivities at the threshold.}
\item{\code{specificity}: The specificities at the threshold.}
\item{\code{ppv}: The positive predictive values at the threshold.}
\item{\code{prevalence}: The prevalence of the endpoint in the study population, as provided in the input.}
}
}
\description{
Diagnostic performance: Positive predictive values (PPV) from a ROC curve
}
