% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-roc.R
\name{ppv.seatsp}
\alias{ppv.seatsp}
\title{Diagnostic performance: Sensitivity from specificity, PPV and prevalence}
\usage{
ppv.seatsp(sp, ppv, prevalence)
}
\arguments{
\item{sp}{Specificity.}

\item{ppv}{Positive predictive value.}

\item{prevalence}{Prevalence of the endpoint in the study population.}
}
\value{
Sensitivity from specificity, PPV and prevalence.
}
\description{
Diagnostic performance: Sensitivity from specificity, positive predictive value (PPV) and prevalence of the endpoint
}
