% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-roc.R
\name{ppv.spatse}
\alias{ppv.spatse}
\title{Diagnostic performance: Specificity from sensitivity, PPV and prevalence}
\usage{
ppv.spatse(se, ppv, prevalence)
}
\arguments{
\item{se}{Sensitivity.}

\item{ppv}{Positive predictive value.}

\item{prevalence}{Prevalence of the endpoint in the study population.}
}
\value{
Specificity from sensitivity, PPV and prevalence.
}
\description{
Diagnostic performance: Specificity from sensitivity, positive predictive value (PPV) and prevalence of the endpoint
}
